feat(api): 重构 API 客户端，添加请求和响应拦截器，统一处理认证和错误
