+++ src/api/apiClient.ts
@@ -1,0 +1,53 @@
+import axios, { type AxiosError, type AxiosResponse, type InternalAxiosRequestConfig } from 'axios';
+
+const apiClient = axios.create({
+  baseURL: import.meta.env.VITE_API_URL || '/api',
+  timeout: 10000,
+  headers: {
+    'Content-Type': 'application/json',
+  },
+});
+
+apiClient.interceptors.request.use(
+  (config: InternalAxiosRequestConfig) => {
+    const token = localStorage.getItem('accessToken');
+    if (token && config.headers) {
+      config.headers.Authorization = `Bearer ${token}`;
+    }
+
+    return config;
+  },
+  (error: AxiosError) => Promise.reject(error)
+);
+
+apiClient.interceptors.response.use(
+  (response: AxiosResponse) => response,
+  (error: AxiosError) => {
+    if (error.response) {
+      const { status } = error.response;
+
+      switch (status) {
+        case 401:
+          localStorage.removeItem('accessToken');
+          window.location.href = '/login';
+          break;
+        case 404:
+          console.error('请求的资源不存在');
+          break;
+        case 500:
+          console.error('服务器内部错误');
+          break;
+        default:
+          console.error(`请求错误: ${status}`);
+      }
+    } else if (error.request) {
+      console.error('网络错误或服务器无响应');
+    } else {
+      console.error('请求配置错误:', error.message);
+    }
+
+    return Promise.reject(error);
+  }
+);
+
+export default apiClient;
